AMQP-294 Remove Class and Package Tangles
Move Consumer Channel Registry methods to new class to avoid
a class tangle between RabbitResourceHolder and ConnectionFactoryUtils.

Move exception translation to a new support class to avoid package
tangle between connection and support.
