+++ spring-rabbit/src/main/java/org/springframework/amqp/rabbit/connection/ConnectionFactoryUtils.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2002-2012 the original author or authors.
+ * Copyright 2002-2013 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
@@ -15,9 +15,8 @@ package org.springframework.amqp.rabbit.connection;
 
 import java.io.IOException;
 
-import org.apache.commons.logging.Log;
-import org.apache.commons.logging.LogFactory;
 import org.springframework.amqp.AmqpIOException;
+import org.springframework.amqp.rabbit.support.ConsumerChannelRegistry;
 import org.springframework.transaction.support.ResourceHolderSynchronization;
 import org.springframework.transaction.support.TransactionSynchronization;
 import org.springframework.transaction.support.TransactionSynchronizationManager;
@@ -38,46 +37,6 @@ import com.rabbitmq.client.Channel;
  * @author Gary Russell
  */
 public class ConnectionFactoryUtils {
-
-	private static final Log logger = LogFactory.getLog(ConnectionFactoryUtils.class);
-
-	private static final ThreadLocal<Channel> consumerChannel = new ThreadLocal<Channel>();
-
-	/**
-	 * If a listener container is configured to use a RabbitTransactionManager, the
-	 * consumer's channel is registered here so that it is used as the bound resource
-	 * when the transaction actually starts. It is normally not necessary to use
-	 * an external transaction manager because local transactions work the same in that
-	 * the channel is bound to the thread. This is for the case when a user happens
-	 * to wire in a RabbitTransactionManager.
-	 * @param channel
-	 */
-	public static void registerConsumerChannel(Channel channel) {
-		if (logger.isDebugEnabled()) {
-			logger.debug("Registering consumer channel" + channel);
-		}
-		consumerChannel.set(channel);
-	}
-
-	/**
-	 * See registerConsumerChannel. This method is called to unregister
-	 * the channel when the consumer exits.
-	 */
-	public static void unRegisterConsumerChannel() {
-		if (logger.isDebugEnabled()) {
-			logger.debug("Unregistering consumer channel" + consumerChannel.get());
-		}
-		consumerChannel.remove();
-	}
-
-	/**
-	 * See registerConsumerChannel. This method is called to retrieve the
-	 * channel for this consumer.
-	 */
-	public static Channel getConsumerChannel() {
-		return consumerChannel.get();
-	}
-
 	/**
 	 * Determine whether the given RabbitMQ Channel is transactional, that is, bound to the current thread by Spring's
 	 * transaction facilities.
@@ -163,7 +122,7 @@ public class ConnectionFactoryUtils {
 				connection = resourceFactory.createConnection();
 				resourceHolderToUse.addConnection(connection);
 			}
-			channel = consumerChannel.get();
+			channel = ConsumerChannelRegistry.getConsumerChannel();
 			if (channel == null) {
 				channel = resourceFactory.createChannel(connection);
 			}
